Convert Observable into a pseudo-abstract class
We need to be able to declare properties like this:

var streamOfStrings: Observable<String> {

}

However, this will not work because Observable does not have a generic type.
We apparently have to use type-erasure in order to make this work. But, the
easy way to it is just make a pseudo-abstract class for now.